requirements.txt
torch
torchvision
numpy
matplotlib
imageio
Pillow
pandas
scikit-learn
